requirements.txt
celery==4.4.7
Django==3.2.6
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2-binary==2.8.6
python-decouple==3.4
python-memcached==1.59
requests==2.26.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
coreapi==2.3.3
django-filter
graphene-django
drf-writable-nested==0.6.3
drf-nested-routers==0.93.4
